Make the timer display the exact time
Previously, the timer only worked correctly starting from the second TOTP generation. When the app was hot restarted or closed and started again, the timer sped up by counting from 30 to the lifespan of the TOTP. Now, the timer and the animation displays the correct time and lifespan of the TOTP code regardless of the lifespan of the app itself.